Add exception and fix quantity calculation bug
If the items in job order is removed, it will not be immediately updated to the database.
The number of part items in the database will be updated when the whole job order is updated.
When add the new item to the job order, the system will compare the number of item with the one in the database.